requirements.txt
# LockSpot Backend - Django Requirements
# Python 3.10+

# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Admin Theme
django-jazzmin>=3.0.0

# Database
mysqlclient>=2.2.0  # For MySQL (optional)

# Authentication
PyJWT>=2.8.0

# Image Processing
Pillow>=10.0.0

# Environment
python-dotenv>=1.0.0

# QR Code Generation
qrcode>=7.4.0

# Production Server
gunicorn>=21.0.0
whitenoise>=6.6.0

# Development
django-debug-toolbar>=4.2.0
